--- a/executor/hash_table_test.go
+++ b/executor/hash_table_test.go
@@ -19,7 +19,7 @@ import (
 )
 
 func (s *pkgTestSuite) TestRowHashMap(c *C) {
-	m := newRowHashMap()
+	m := newRowHashMap(0)
 	m.Put(1, chunk.RowPtr{ChkIdx: 1, RowIdx: 1})
 	c.Check(m.Get(1), DeepEquals, []chunk.RowPtr{{ChkIdx: 1, RowIdx: 1}})
 
@@ -29,7 +29,7 @@ func (s *pkgTestSuite) TestRowHashMap(c *C) {
 			rawData[i] = append(rawData[i], chunk.RowPtr{ChkIdx: uint32(i), RowIdx: uint32(j)})
 		}
 	}
-	m = newRowHashMap()
+	m = newRowHashMap(0)
 	// put all rawData into m vertically
 	for j := uint64(0); j < initialEntrySliceLen*9; j++ {
 		for i := 9; i >= 0; i-- {