<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Part of #11607
It reorgs codes for the hash table and introduces `hashRowContainer`. So that In further pull request, I can implement `spillOutToDisk` feature for `hashRowContainer`

### What is changed and how it works?

- [x] reorg codes for hash table in `HashJoinExec`
- [x] make a suitable `maxEntrySliceLen`, it's currently too large(8k) for OLTP queries.
- [x] add unit test for `rowHashMap`.
- [x] try to replace FNV64 to another hash function. (abandon, FNV64 is the fastest.)
- [x] initialize the size of `rowHashMap` from the estimated row count for better performance.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test

Code changes

 - None

Side effects

 - None

Related changes

 - None

Release note

 - None

### Benchmark

```
name                                                                     old time/op    new time/op    delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-12              760ms ± 6%     781ms ± 7%     ~     (p=0.421 n=5+5)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-12                168ms ± 5%     166ms ± 2%     ~     (p=0.841 n=5+5)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-12     512ms ± 0%     544ms ± 1%   +6.19%  (p=0.016 n=4+5)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-12      12.2ms ± 2%    12.5ms ± 8%     ~     (p=0.548 n=5+5)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-12          2.46µs ± 2%    2.94µs ± 0%  +19.46%  (p=0.008 n=5+5)

name                                                                     old alloc/op   new alloc/op   delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-12              305MB ± 0%     304MB ± 0%   -0.28%  (p=0.008 n=5+5)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-12                305MB ± 0%     304MB ± 0%   -0.28%  (p=0.008 n=5+5)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-12    7.97MB ± 0%    7.11MB ± 0%  -10.80%  (p=0.008 n=5+5)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-12      7.97MB ± 0%    7.11MB ± 0%  -10.80%  (p=0.008 n=5+5)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-12          1.62kB ± 0%    1.96kB ± 0%  +20.74%  (p=0.008 n=5+5)

name                                                                     old allocs/op  new allocs/op  delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-12               306k ± 0%      306k ± 0%   -0.08%  (p=0.008 n=5+5)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-12                 306k ± 0%      306k ± 0%   -0.09%  (p=0.008 n=5+5)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-12     4.07k ± 2%     3.80k ± 1%   -6.56%  (p=0.008 n=5+5)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-12       4.04k ± 0%     3.79k ± 0%   -6.41%  (p=0.008 n=5+5)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-12            9.00 ± 0%     14.00 ± 0%  +55.56%  (p=0.008 n=5+5)
```
[old.txt](https://github.com/pingcap/tidb/files/3559836/old.txt)
[new.txt](https://github.com/pingcap/tidb/files/3559840/new.txt)